Upgrade audio player with custom controls and styling
The commit introduces changes to both the audio player's styling and control functionality. It shifts from utilizing the default browser player controls to a custom set of controls, offering a better look and user experience. On the HTML side, instead of using webkit-media-controls, a custom structure has been used for play control, volume control and a navigation bar.

Further, changes in the CSS file allow the audio player to have a distinct and attractive look. JavaScript file changes allow these custom controls to function as intended, this includes playing and pausing a song, adjusting volume and time navigation.

These changes allow for a more aesthetically pleasing user interface and a smoother user experience.