Refactor volume control in music player
This commit refactors the volume control mechanism in the music player on our site. We have switched the volume control from a click event-driven vertical slider to a more standard horizontal input slider, making it more intuitive for users.

In addition, redundant code concerning the old volume control method was removed. This code refactoring not only enhances the usability of the music player but also improves the readability and maintainability of our JavaScript code.